# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PayDuka PoS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Product {
  id: number;
  name: string;
  sku: string;
  category: string;
  price: number;
  stock: number;
  lowStockThreshold: number;
  unit: string;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProductBody {
  name: string;
  sku: string;
  category: string;
  price: number;
  stock: number;
  lowStockThreshold: number;
  unit: string;
  /** @nullable */
  imageUrl?: string | null;
}

export interface UpdateProductBody {
  name?: string;
  sku?: string;
  category?: string;
  price?: number;
  stock?: number;
  lowStockThreshold?: number;
  unit?: string;
  /** @nullable */
  imageUrl?: string | null;
}

export interface SaleItem {
  id: number;
  saleId: number;
  productId: number;
  productName: string;
  quantity: number;
  unitPrice: number;
  subtotal: number;
}

export interface Sale {
  id: number;
  total: number;
  paymentMethod: string;
  /** @nullable */
  customerName?: string | null;
  pdukaEarned: number;
  /** @nullable */
  txHash?: string | null;
  createdAt: string;
}

export interface SaleWithItems {
  id: number;
  total: number;
  paymentMethod: string;
  /** @nullable */
  customerName?: string | null;
  pdukaEarned: number;
  /** @nullable */
  txHash?: string | null;
  createdAt: string;
  items: SaleItem[];
}

export interface CreateSaleItemInput {
  productId: number;
  quantity: number;
}

export type CreateSaleBodyPaymentMethod =
  (typeof CreateSaleBodyPaymentMethod)[keyof typeof CreateSaleBodyPaymentMethod];

export const CreateSaleBodyPaymentMethod = {
  cash: "cash",
  usdc: "usdc",
  pol: "pol",
  pduka: "pduka",
  card: "card",
} as const;

export interface CreateSaleBody {
  items: CreateSaleItemInput[];
  paymentMethod: CreateSaleBodyPaymentMethod;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  txHash?: string | null;
}

export interface DashboardSummary {
  todayRevenue: number;
  todaySalesCount: number;
  todayPdukaEarned: number;
  yesterdayRevenue: number;
  yesterdaySalesCount: number;
  revenueChangePercent: number;
  salesCountChangePercent: number;
  totalProducts: number;
  lowStockCount: number;
  outOfStockCount: number;
}

export interface DailySalesPoint {
  date: string;
  revenue: number;
  salesCount: number;
}

export interface MonthlyComparison {
  currentMonth: string;
  currentRevenue: number;
  currentSalesCount: number;
  previousMonth: string;
  previousRevenue: number;
  previousSalesCount: number;
  revenueChangePercent: number;
  salesCountChangePercent: number;
}

export interface TopProduct {
  productId: number;
  name: string;
  totalRevenue: number;
  totalQuantity: number;
}

export type StockAlertStatus =
  (typeof StockAlertStatus)[keyof typeof StockAlertStatus];

export const StockAlertStatus = {
  low: "low",
  out_of_stock: "out_of_stock",
} as const;

export interface StockAlert {
  id: number;
  name: string;
  sku: string;
  stock: number;
  lowStockThreshold: number;
  status: StockAlertStatus;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export type ListSalesParams = {
  limit?: number;
};