# lib/api-zod/src/generated/types/createProductBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PayDuka PoS API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductBody {
  name: string;
  sku: string;
  category: string;
  price: number;
  stock: number;
  lowStockThreshold: number;
  unit: string;
  /** @nullable */
  imageUrl?: string | null;
}