# lib/api-zod/src/generated/types/createSaleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PayDuka PoS API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateSaleBodyPaymentMethod } from "./createSaleBodyPaymentMethod";
import type { CreateSaleItemInput } from "./createSaleItemInput";

export interface CreateSaleBody {
  items: CreateSaleItemInput[];
  paymentMethod: CreateSaleBodyPaymentMethod;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  txHash?: string | null;
}