# lib/api-zod/src/generated/types/dailySalesPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PayDuka PoS API
 * OpenAPI spec version: 0.1.0
 */

export interface DailySalesPoint {
  date: string;
  revenue: number;
  salesCount: number;
}