# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PayDuka PoS API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  todayRevenue: number;
  todaySalesCount: number;
  todayPdukaEarned: number;
  yesterdayRevenue: number;
  yesterdaySalesCount: number;
  revenueChangePercent: number;
  salesCountChangePercent: number;
  totalProducts: number;
  lowStockCount: number;
  outOfStockCount: number;
}